enhancements(utils): Merge helpers into utils
There is no need to keep them separate as the serve exactly the same purpose in the end
